✨ feat(restart): train_nnsk
decoupling init data and model init, adding the restart module

several bugs in strain mode is fixed. Including the error when anglar  m of first atom is less than the second; bugs in load params and hamil_sk. A problem exist that restart now is not able to generate the anglr m change and onsite mode change, since the initialization of optimier requires a state_dict. Where a new param reload method need to be wrote. Also, device and atom type is not well used in a lot of module, this should be fixed later if moving to GPU.