refactor(data preprocess): remove the cut off options from info.json (#200)
* refactor(data preprocess): remove the cut off options from info.json and collect the values from input.json

* update LMDB info.json. not need anymore.

* refactor(default_dataset): refactor the _TrajData for ase data.

Previous the ase data will be transferred into text file and then loaded by the _TrajData. now i refactor the function.

both text and ase data are treated equally. will works as a class funtion to initial the _TrajData class.

* add print logo in main and format some of  the logger.info

* update argcheck  collect_cutoffs.  add new function with  get_cutoffs_from_model_options .

* Fix(get_cutoffs_from_model_options) : fix rcut in  powerlaw and varTang96.

For powerlaw and varTang96, the rs is not exactly the hard cutoff. so when extract the r_max for data. we have to use rs + 5 * w; but for other method just use rs.

* update band post process.

* update test

* update test

* update build and get_cutoffs_from_model_options to support the rmax to be dict.

* refactor(build dataset): change build_dataset from function to a class instance and add from_model class function.

note, compared to the previous build_dataset, this one is more flexible.
previous build_dataset is a function. now i define a class DataBuilder and re-defined __call__ function.  then build_dataset is an instance of DataBuilder class. so i can use build_dataset.from_model() to build dataset from model. at the same time the previous way to use  build_dataset is still available. like build_dataset(...).

* add checkcutoff in dataset builder.

* update AtomicData_options to make it compatible with older versions

* Update argcheck.py

---------

Co-authored-by: Yinzhanghao Zhou <64253517+floatingCatty@users.noreply.github.com>